Weekend progress

+++ actions/delete.php
@@ -1,5 +1,10 @@
 <?php
+
+	session_start();
+	if($_SESSION['level']!="staff") die();
+
 	include_once("../connections/connect.php");
+
 	$sql = "DELETE FROM material WHERE id = '".$_POST["id"]."'";
 	if(pg_query($db, $sql))
 	{
